[back] move : contest/controller/service -> controller와 서비스 폴더 분리
